refactor Root component
